III-578: refactoring LDProjector tests
