+++ test/Place/ReadModel/JSONLD/place_with_image.cdbxml.xml
