+++ test/Place/ReadModel/JSONLD/place_with_short_description.cdbxml.xml
