+++ test/Place/ReadModel/JSONLD/place_without_image.cdbxml.xml
